Do not use CS6 as the default packet marking. Enable ECN.
CS6 is mistreated by many wireless-n stacks to put packets into the VO
queue. I have long had it enabled and have been observing queuing
problems and packet loss that I'd like to address by disabling it.

In the long run, I could see restoring CS6 for the most important
use case - network alerts - but not for all babel traffic as has been
the case for years now.

ECN is helpful for me in tracking the effects of the various fq_codel
derived algorithms in determining the difference between native packet
loss (due to interference) and congestive packet loss. Not doing
anything intelligent with it is not a problem for my limited
distributions.