Hopefully atomic route updates
This essentially

the trick was to just give up on the new KERNEL_INFINITY setting
(which shifts things to localhost on linux systems), and flush
add that, while retaining add/flush semantics with a new metric.

Cost: doubles the number of route_modify related syscalls.

Still wanted to somehow halve them.....