Atomic route switches finally
I would not trust this code yet, but it does (finally) appear to be
doing the right things to do an atomic route modify in place in
all the circumstances I've been able to test thus far.

It's ugly as sin, also. I'd like to clean up the netlink stuff in
general.

The code still prints a lot of errors while flushing routes on
interfaces that go down. I assume this is because the kernel flushed
them for us, but adding better error checking here would be good.

Other annoyances are that the ipv4 dhcp acquired routes nowadays
get a metric of 600 (which babel overrides), but the dhpv6 acquired
addresses got 600 (which babel is less than, as it defaults to 1024.

RA inserts a metric of 1024. I have source specific routing in place,
so I'd hope that in case of equivalent metrics, the source specific
default route would win.

I think modifying the default rabel behavior here to specify a metric of
1 for ipv6 would probably help. Also making babel be more aggressive
(optionally) about deleting routes it sees from other daemons might
help.