Update makefile to use -O3 and whole program optimization
With -Os compiling a whole program saves 2k on the object,
down to 70k on arm.

With -O3, compile the whole program is about 4k less (110Kbytes)
and I have this delusional hope that the optimizer found more
opportunities to optimize.

In earlier passes I tried to identify pre-existing hotspots
and inline them by the general rule that < 4 invocations
is reasonable to inline. Particularly in the pointer-walking
code, which tended to reformat the args 3 times before doing
useful work, it is my hope that this vastly improved registerization
on arches that have plenty of registers. Also everybody has an
amazingly fast mem* set of functions nowadays, this exposes enough
info for that to always happen.

That said - benchmarking awaits! (after the kernel bug is resolved)