And finally... bottlenecking on solving bf I think.
We need to periodically check during the core loop here to see
if we've run out of compute time, and make sure we get a hello out.

Otherwise, we drop off the net. Which causes catastrophic results
for myself and everyone else.

My guess is about every 512 passes through the loop, make sure we're
not out of time.