I shrunk the proto and later found out it was used for something else
so either we can use up an existing proto or go back to a short here.
But my goal was to find the darn corruption error that is still
plaguing the entire codebase, and more than fixing this is
inspecting netlink for size_t stuff, trying to figure out why
we get martian routes in the first place, and so on.