Further neon-optimize the compare
This version avoids reloading the essentially constant compared
prefixes. Perhaps I can convince the compiler to be smarter about
emitting this rather than coding it up directly.

It still stalls out enormously, even with the prefetch.

Now... it turns as to why this routine is dominating the runtime
is because xroute is using a bubble sort to go through the kernel
routes. Replacing that with a sorted list will improve matters
enormously.

And most people are not as crazy as I am in dumping so many
routes into the daemon, in a test.