Some shims for the adapteva epiphany 16 core co-processor
I had *most* of this working 2 years ago, with a good set of shims,
and was basically only lacking the basic kernel netlink to epiphany
code, as well as malloc in the epiphany was buggy, as well as other
issues.

I don't know where all that work went at the moment, but I got a
bug up my posterior to at least, this time, check some of the work
in so I can find it again, because the epiphany SDK and compiler have
got a LOT better. I do wish they ported musl rather than newlib...

It is totally not an ideal co-processor for this sort of thing,
but the core ideas behind what went into porting this also apply to the
riscv-128 work also lying around, and sort of have been driving my
ideas for modifying babel to work better on 128 bit registers, so...