Restore v4mapped comparison to prefix_cmp
I had ripped out an earlier commit by Mattieu because I didn't
understand why it was in there - and the overhead on computing
it was enormous, and right in the hot path.

I am still not happy about it, and will think about how to
do things better:

From the prior commit:

When the v6 prefix is ::/0, then it is considered as less specific than
any v4-mapped prefix.

Remark that "no source prefix" is encoded by "src_plen == 0".  This
might be an issue, since "src_plen == 0" means ::/0, a v6 prefix.  So
the route (1.1.1.1/32 from 0.0.0.0/0) is internally represented by
(1.1.1.1/32 from ::/0).  When compared with  (::/0 from 2001:db8::/32),
the two source prefixes are comparable.  This is not a problem in
practice, since 2 pairs of prefixes are disjoint whether the dst
prefixes or the src prefixes are disjoint.