Do not know the difference between RTNH_F_ONLINK and not
I had a bit of hope that suppressing it might do something
"interesting". It didn't. Not that I could tell.

Checking this in so I'll remember later.

Also found that ENODEV is a valid error message from the kernel.

Which I'm not checking for and so far as I know have never seen.